possible fix for float issue

+++ x/structs/keeper/player.go
@@ -217,7 +217,7 @@ func (k Keeper) GetPlayerCharge(ctx context.Context, playerId string) (charge ui
     lastActionBlock := k.GetGridAttribute(ctx, GetGridAttributeIDByObjectId(types.GridAttributeType_lastAction, playerId))
     blockSpan := uint64(ctxSDK.BlockHeight()) - lastActionBlock
 
-    result := types.Charge_Volts * (1 - math.Pow(math.Exp(1), -(blockSpan/(types.Charge_Resistance*types.Charge_Capacitance))))
+    result := types.Charge_Volts * (1 - math.Pow(math.Exp(1), -(float64(blockSpan)/(types.Charge_Resistance*types.Charge_Capacitance))))
     charge = uint64(result)
 
 	return
